file & structure refactor with/for redux setp attempt 1, Note: Needs some more work
