+++ package.json
@@ -12,6 +12,7 @@
     "@types/react-redux": "^7.1.4",
     "@types/react-router": "^5.1.1",
     "@types/react-router-dom": "^5.1.0",
+    "axios": "^0.19.0",
     "react": "^16.10.1",
     "react-dom": "^16.10.1",
     "react-redux": "^7.1.1",
