+++ src/store/configureStore.ts
@@ -1,6 +1,7 @@
-import { createStore, applyMiddleware, combineReducers } from 'redux';
-import thunk from 'redux-thunk';
-import userReducer from '../reducers/user';
+import { applyMiddleware, combineReducers, createStore } from 'redux'
+import thunk from 'redux-thunk'
+
+import userReducer from '../reducers/user.reducer'
 
 export const rootReducer = combineReducers({
     user: userReducer
